refactor(compactor): move merge batch planning into merge/plan.rs
